[D83T-46] 충돌 해결

+++ src/main/java/d83t/bpmbackend/domain/aggregate/location/repository/LocationRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface LocationRepository extends JpaRepository<Location, Long> {
-    Optional<Location> findByLatitudeAndLongitude(int latitude, int longitude);
+    Optional<Location> findByLatitudeAndLongitude(double latitude, double longitude);
 }
